% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data_consistency}
\alias{check_data_consistency}
\title{Check Data Consistency}
\usage{
check_data_consistency(time_points, surv_probs)
}
\arguments{
\item{time_points}{Vector of time points from digitized curve}

\item{surv_probs}{Vector of survival probabilities from digitized curve}
}
\value{
List with validation results and messages
}
\description{
This function checks if digitized survival data is consistent and valid.
}
\examples{
# Check data consistency
time_points <- c(0, 6, 12, 18, 24)
surv_probs <- c(1, 0.9, 0.8, 0.75, 0.7)
check_data_consistency(time_points, surv_probs)
}
