% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_time_units}
\alias{convert_time_units}
\title{Convert Between Time Units}
\usage{
convert_time_units(time, from_unit, to_unit)
}
\arguments{
\item{time}{Numeric vector of time values to convert}

\item{from_unit}{Original time unit ("days", "weeks", "months", "years")}

\item{to_unit}{Target time unit ("days", "weeks", "months", "years")}
}
\value{
Numeric vector of converted time values
}
\description{
This function converts time values between different units,
which can be useful when dealing with curves using different time scales.
}
\examples{
# Convert 24 months to years
convert_time_units(24, from_unit = "months", to_unit = "years")

# Convert 5 years to months
convert_time_units(5, from_unit = "years", to_unit = "months")
}
