% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_total_events}
\alias{estimate_total_events}
\title{Estimate Total Events from Survival Curve}
\usage{
estimate_total_events(time_points, surv_probs, n_risk_initial)
}
\arguments{
\item{time_points}{Vector of time points from digitized curve}

\item{surv_probs}{Vector of survival probabilities from digitized curve}

\item{n_risk_initial}{Initial number at risk}
}
\value{
Estimated total number of events
}
\description{
This function estimates the total number of events from a digitized survival curve
and initial number at risk.
}
